Refactor #578: admin/handler_stubs.go を完全消滅させ thematic 14 ファイルへ分解
#574 / #575 / #576 / #577 で確立した handler_stubs 解体パターンの最終段。
admin package に残っていた handler_stubs.go (1313 行) と handler_stubs_test.go
(687 行) を完全削除し、handler を 14 のテーマ別ファイル + email.go 拡張
に再配置する。pure file move + helper organization のみで挙動変更なし。

handler 新規 / 拡張 (15 ファイル):
- forward_abuse_report.go: ForwardAbuseUserReport
- server_stats.go:        GetIndexStats / GetTableStats / GetUserIPs / ServerInfo
- password_reset.go:      ResetPassword + sendPasswordResetEmail / issueTemporaryPassword
- email.go (拡張):         SendEmail (既存 sendWebhookTest と同居、sendWebhookTest
                          の system_webhook.go 移動は #583 残作業)
- proxy_account.go:       UpdateProxyAccount
- drive.go:               DriveCleanRemoteFiles / DriveCleanup / DriveFiles /
                          DriveShowFile + packDriveFileAdmin
- emoji.go:               9 Emoji* handlers + dedupe
- captcha.go:             CaptchaCurrent / CaptchaSave
- ad.go:                  Ad* (4) + millisOrNow / millisOrDefault
- avatar_decorations.go:  AvatarDecorations* (4)
- federation_admin.go:    Federation* (4)
- invite.go:              InviteCreate / InviteList + packInviteTickets
- promo.go:               PromoCreate
- relays.go:              Relays* (3)

test 新規 / 拡張:
- handler_test.go:        adminUser / assertError / TestSet* helpers を集約
- forward_abuse_report_test.go (new):  3 tests + stubAbuseForwarder
- server_stats_test.go (new):          6 tests + stubIPRepo
- password_reset_test.go (new):        7 tests + stubPasswordResetRepo / capturedEmail
- email_test.go (new):                 1 test
- proxy_account_test.go (new):         placeholder (詳細は accounts_test.go)
- captcha_test.go (new):               2 thin + 4 detailed (詳細は queue_stats_test.go から移動 — #583 半分回収)
- federation_admin_test.go (new):      6 tests + stubInstanceMetadataFetcher
- relays_test.go (new):                8 tests + fakeRelaySvc
- ad_invite_promo_test.go (拡張):       11 thin nil-repo tests を末尾追加
- drive_emoji_test.go (拡張):           13 thin + 6 EmojiImportZip 詳細 + stubEmojiImportEnqueuer
- queue_stats_test.go:                 captcha 4 テスト削除、Queue* thin 12 tests 追加
- system_webhook_test.go:              SystemWebhook* thin 6 tests 追加
- abuse_report_notification_test.go:   AbuseReportNotificationRecipient* thin 5 tests 追加

#583 の半分 (queue_stats_test.go の captcha 混入解消) を本 PR で回収。
sendWebhookTest を system_webhook.go へ移す残作業は引き続き #583 で扱う。

Coverage: 84.0% (develop と完全一致、admin 閾値 80% を上回る)。
test 関数数: 317 (develop と完全一致)。

これで mk-go の API package から *handler_stubs*.go ファイルが完全消滅。

Closes #578
