--- internal/api/admin/handler_stubs.go
@@ -1,1313 +1,0 @@
-package admin
-
-import (
-	"crypto/rand"
-	"encoding/hex"
-	"log/slog"
-	"net/http"
-	"time"
-
-	"github.com/labstack/echo/v4"
-	"github.com/shiroha-a/mk/internal/api/apierr"
-	"github.com/shiroha-a/mk/internal/core/serverstats"
-	"github.com/shiroha-a/mk/internal/entity"
-	"github.com/shiroha-a/mk/internal/misc"
-	"github.com/shiroha-a/mk/internal/misc/smtp"
-	"github.com/shiroha-a/mk/internal/model"
-	"github.com/shiroha-a/mk/internal/queue"
-	"github.com/shiroha-a/mk/internal/server/middleware"
-	"golang.org/x/crypto/bcrypt"
-)
-
-// AccountsDelete / AccountsFindByEmail / DeleteAccount /
-// scheduleAccountCascade は accounts.go に移動済 (#574)。
-// DeleteAllFilesOfUser は moderation.go に移動済 (#574)。
-
-// ForwardAbuseUserReport handles POST /api/admin/forward-abuse-user-report.
-//
-// 対象ユーザーがリモートの場合、system actor 署名で origin インスタンス
-// の inbox へ ActivityPub Flag を配送し、DB 側 forwarded=true も立てる。
-// ローカル通報の場合は配送スキップで DB フラグのみ更新する。
-func (h *Handler) ForwardAbuseUserReport(c echo.Context) error {
-	var req struct {
-		ReportID string `json:"reportId"`
-	}
-	_ = c.Bind(&req)
-	if req.ReportID == "" {
-		return c.NoContent(http.StatusNoContent)
-	}
-	if h.abuseForwarder != nil {
-		if err := h.abuseForwarder.ForwardReport(req.ReportID); err != nil {
-			return apierr.JSONInternalError(c)
-		}
-		return c.NoContent(http.StatusNoContent)
-	}
-	// forwarder 未配線時のフォールバック: DB フラグだけ更新する (テストや
-	// federation stack 未初期化パスで有効)。
-	if h.abuseRepo != nil {
-		_ = h.abuseRepo.UpdateFields(req.ReportID, map[string]any{"forwarded": true})
-	}
-	return c.NoContent(http.StatusNoContent)
-}
-
-// GetIndexStats handles POST /api/admin/get-index-stats.
-//
-// Returns per-index row counts from pg_stat_user_indexes so the admin UI can
-// spot hot or unused indexes.
-func (h *Handler) GetIndexStats(c echo.Context) error {
-	if h.adminDB == nil {
-		return c.JSON(http.StatusOK, []any{})
-	}
-	type row struct {
-		Relname      string `json:"tablename" gorm:"column:relname"`
-		Indexrelname string `json:"indexname" gorm:"column:indexrelname"`
-		IdxScan      int64  `json:"idx_scan" gorm:"column:idx_scan"`
-		IdxTupRead   int64  `json:"idx_tup_read" gorm:"column:idx_tup_read"`
-	}
-	var rows []row
-	if err := h.adminDB.Raw(`
-		SELECT relname, indexrelname, idx_scan, idx_tup_read
-		FROM pg_stat_user_indexes
-		ORDER BY relname, indexrelname
-	`).Scan(&rows).Error; err != nil {
-		return c.JSON(http.StatusInternalServerError, apierr.Error("INTERNAL_ERROR", "Internal error.", "00000000-0000-0000-0000-000000000000"))
-	}
-	return c.JSON(http.StatusOK, rows)
-}
-
-// GetTableStats handles POST /api/admin/get-table-stats.
-//
-// Returns per-table size and row estimate via pg_stat_user_tables joined with
-// pg_relation_size for quick capacity planning.
-func (h *Handler) GetTableStats(c echo.Context) error {
-	if h.adminDB == nil {
-		return c.JSON(http.StatusOK, map[string]any{})
-	}
-	type row struct {
-		Relname  string `gorm:"column:relname"`
-		Count    int64  `gorm:"column:row_count"`
-		SizeBase int64  `gorm:"column:size_base"`
-		SizeIdx  int64  `gorm:"column:size_idx"`
-	}
-	var rows []row
-	if err := h.adminDB.Raw(`
-		SELECT c.relname,
-		       COALESCE(s.n_live_tup, 0) AS row_count,
-		       pg_relation_size(c.oid) AS size_base,
-		       COALESCE(pg_indexes_size(c.oid), 0) AS size_idx
-		FROM pg_class c
-		LEFT JOIN pg_stat_user_tables s ON s.relid = c.oid
-		WHERE c.relkind = 'r'
-		  AND c.relnamespace IN (SELECT oid FROM pg_namespace WHERE nspname = 'public')
-		ORDER BY c.relname
-	`).Scan(&rows).Error; err != nil {
-		return c.JSON(http.StatusInternalServerError, apierr.Error("INTERNAL_ERROR", "Internal error.", "00000000-0000-0000-0000-000000000000"))
-	}
-	// Misskey 本家は { tableName: { count, size } } の map 形式で返すのでそれに合わせる。
-	result := make(map[string]any, len(rows))
-	for _, r := range rows {
-		result[r.Relname] = map[string]any{
-			"count": r.Count,
-			"size":  r.SizeBase + r.SizeIdx,
-		}
-	}
-	return c.JSON(http.StatusOK, result)
-}
-
-// GetUserIPs handles POST /api/admin/get-user-ips.
-func (h *Handler) GetUserIPs(c echo.Context) error {
-	if h.userIPRepo == nil {
-		return c.JSON(http.StatusOK, []any{})
-	}
-	var req struct {
-		UserID string `json:"userId"`
-	}
-	if err := c.Bind(&req); err != nil || req.UserID == "" {
-		return c.JSON(http.StatusBadRequest, apierr.Error("INVALID_PARAM", "userId is required.", "3d81ceae-475f-4600-b2a8-2bc116157532"))
-	}
-	ips, err := h.userIPRepo.ListByUser(req.UserID, 30)
-	if err != nil {
-		return c.JSON(http.StatusOK, []any{})
-	}
-	result := make([]map[string]any, 0, len(ips))
-	for _, ip := range ips {
-		result = append(result, map[string]any{
-			"ip":        ip.IP,
-			"createdAt": ip.CreatedAt.UTC().Format("2006-01-02T15:04:05.000Z"),
-		})
-	}
-	return c.JSON(http.StatusOK, result)
-}
-
-// ResetPassword handles POST /api/admin/reset-password.
-//
-// 優先: 対象ユーザーに verified email があれば reset token を発行してメール
-// リンクを送り、HTTP 200 で {"sent": true} を返す (ユーザーが自分で新パス
-// ワードを設定する本家互換フロー #186)。
-//
-// Fallback: email が verify されていない、repo / sender が未配線、
-// etc. の場合は従来どおりランダムなテンポラリパスワードを生成してその場で
-// 更新し、{"password": "..."} を返す。旧管理 UI との互換を保つ最後の砦。
-func (h *Handler) ResetPassword(c echo.Context) error {
-	var req struct {
-		UserID string `json:"userId"`
-	}
-	if err := c.Bind(&req); err != nil || req.UserID == "" {
-		return c.JSON(http.StatusBadRequest, apierr.Error("INVALID_PARAM", "userId is required.", "3d81ceae-475f-4600-b2a8-2bc116157532"))
-	}
-	if sent := h.sendPasswordResetEmail(req.UserID); sent {
-		return c.JSON(http.StatusOK, map[string]any{"sent": true})
-	}
-	return h.issueTemporaryPassword(c, req.UserID)
-}
-
-// sendPasswordResetEmail attempts the token+email flow. Returns true only
-// when a reset token was persisted and the email handed off to the sender.
-// Any missing dependency / unverified email triggers false for the caller
-// to pick up the legacy path.
-func (h *Handler) sendPasswordResetEmail(userID string) bool {
-	if h.resetReqRepo == nil || h.emailSender == nil || h.userRepo == nil {
-		return false
-	}
-	profile, err := h.userRepo.FindProfileByUserID(userID)
-	if err != nil || profile.Email == nil || *profile.Email == "" || !profile.EmailVerified {
-		return false
-	}
-	token := misc.SecureRandomHex(64)
-	now := time.Now()
-	resetReq := &model.PasswordResetRequest{
-		ID:     h.idGen.Generate(now),
-		Token:  token,
-		UserID: userID,
-	}
-	if err := h.resetReqRepo.Create(resetReq); err != nil {
-		return false
-	}
-	link := h.serverURL + "/reset-password/" + token
-	go h.emailSender(*profile.Email, "Password reset",
-		"An administrator initiated a password reset for your account.\n"+
-			"Use the following link within 30 minutes to set a new password:\n"+link)
-	return true
-}
-
-// issueTemporaryPassword is the legacy path retained for installations where
-// email is not configured or the user has no verified address. Generates a
-// random password, updates the profile, and returns it to the admin.
-func (h *Handler) issueTemporaryPassword(c echo.Context, userID string) error {
-	b := make([]byte, 8)
-	_, _ = rand.Read(b)
-	newPass := hex.EncodeToString(b)
-	hash, _ := bcrypt.GenerateFromPassword([]byte(newPass), bcrypt.DefaultCost)
-	if h.userRepo != nil {
-		_ = h.userRepo.UpdateProfile(userID, map[string]any{"password": string(hash)})
-	}
-	return c.JSON(http.StatusOK, map[string]any{"password": newPass})
-}
-
-// SendEmail handles POST /api/admin/send-email.
-func (h *Handler) SendEmail(c echo.Context) error {
-	var req struct {
-		To      string `json:"to"`
-		Subject string `json:"subject"`
-		Text    string `json:"text"`
-	}
-	if err := c.Bind(&req); err != nil || req.To == "" {
-		return c.NoContent(http.StatusNoContent)
-	}
-	// SMTP送信
-	if h.metaRepo != nil {
-		m, err := h.metaRepo.Fetch()
-		if err == nil && m.EnableEmail && m.SmtpHost != nil && m.Email != nil {
-			port := 587
-			if m.SmtpPort != nil {
-				port = *m.SmtpPort
-			}
-			go smtp.SendWithOptions(*m.SmtpHost, port, m.SmtpUser, m.SmtpPass, *m.Email, req.To, req.Subject, req.Text, smtp.Options{ProxyURL: h.smtpProxyURL})
-		}
-	}
-	return c.NoContent(http.StatusNoContent)
-}
-
-// ServerInfo handles POST /api/admin/server-info.
-// meta.enableServerMachineStats が false ならゼロ値を返す。
-func (h *Handler) ServerInfo(c echo.Context) error {
-	if m, err := h.metaRepo.Fetch(); err == nil && m.EnableServerMachineStats {
-		return c.JSON(http.StatusOK, serverstats.Collect())
-	}
-	return c.JSON(http.StatusOK, serverstats.Empty())
-}
-
-// UnsetUserAvatar / UnsetUserBanner / UpdateAbuseUserReport は
-// moderation.go に移動済 (#574)。
-
-// UpdateProxyAccount handles POST /api/admin/update-proxy-account.
-//
-// 本家 update-proxy-account.ts と同じく proxy system user の profile.description
-// を更新する handler (旧 stub は誤って meta.proxyAccountId を書き換えていた、#348)。
-// frontend admin/settings 画面の proxyAccountForm が `{ description }` 形式で
-// 呼び出すのでこれに合わせる。Response は UserDetailed (mk-go では UserLite +
-// description で代替)。
-func (h *Handler) UpdateProxyAccount(c echo.Context) error {
-	var req struct {
-		Description *string `json:"description"`
-	}
-	if err := c.Bind(&req); err != nil {
-		return c.JSON(http.StatusBadRequest, apierr.Error("INVALID_PARAM", "Invalid parameters.", "00000000-0000-0000-0000-000000000000"))
-	}
-	if h.systemAccountFetcher == nil || h.userRepo == nil {
-		return c.NoContent(http.StatusNoContent)
-	}
-	proxy, err := h.systemAccountFetcher.Fetch("proxy")
-	if err != nil || proxy == nil {
-		return c.JSON(http.StatusInternalServerError, apierr.Error("INTERNAL_ERROR", "Internal error.", "00000000-0000-0000-0000-000000000000"))
-	}
-	if req.Description != nil {
-		if err := h.userRepo.UpdateProfile(proxy.ID, map[string]any{"description": req.Description}); err != nil {
-			return c.JSON(http.StatusInternalServerError, apierr.Error("INTERNAL_ERROR", "Internal error.", "00000000-0000-0000-0000-000000000000"))
-		}
-	}
-	// 更新後 profile を再取得して UserDetailed を返す。
-	profile, _ := h.userRepo.FindProfileByUserID(proxy.ID)
-	return c.JSON(http.StatusOK, entity.PackUserDetailed(proxy, profile))
-}
-
-// UpdateUserNote は moderation.go に移動済 (#574)。
-
-// --- drive ---
-
-// DriveCleanRemoteFiles handles POST /api/admin/drive/clean-remote-files.
-func (h *Handler) DriveCleanRemoteFiles(c echo.Context) error {
-	// 単一 DELETE 文なので同期実行で十分。将来バッチ化が必要ならここを
-	// queue ジョブに差し替える。
-	if h.driveFileRepo != nil {
-		_, _ = h.driveFileRepo.DeleteRemoteCache()
-	}
-	return c.NoContent(http.StatusNoContent)
-}
-
-// DriveCleanup handles POST /api/admin/drive/cleanup.
-func (h *Handler) DriveCleanup(c echo.Context) error {
-	if h.driveFileRepo != nil {
-		_, _ = h.driveFileRepo.DeleteOrphans()
-	}
-	return c.NoContent(http.StatusNoContent)
-}
-
-// DriveFiles handles POST /api/admin/drive/files.
-func (h *Handler) DriveFiles(c echo.Context) error {
-	if h.driveFileRepo == nil {
-		return c.JSON(http.StatusOK, []any{})
-	}
-	var req struct {
-		Limit    int    `json:"limit"`
-		SinceID  string `json:"sinceId"`
-		UntilID  string `json:"untilId"`
-		UserID   string `json:"userId"`
-		Origin   string `json:"origin"`
-		Host     string `json:"host"`
-		Hostname string `json:"hostname"`
-		Type     string `json:"type"`
-	}
-	_ = c.Bind(&req)
-	if req.Limit <= 0 || req.Limit > 100 {
-		req.Limit = 30
-	}
-	switch req.Origin {
-	case "", "combined", "local", "remote":
-	default:
-		req.Origin = "combined"
-	}
-	// upstream は host も hostname もどちらも受けて同じ意味で扱う。
-	// frontend (admin/instance-info) が hostname で投げるため互換上両方サポート。
-	host := req.Host
-	if host == "" {
-		host = req.Hostname
-	}
-	files, err := h.driveFileRepo.ListForAdmin(req.UserID, req.Origin, host, req.Type, req.UntilID, req.SinceID, req.Limit)
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, apierr.Error("INTERNAL_ERROR", "Internal error.", "00000000-0000-0000-0000-000000000000"))
-	}
-	out := make([]entity.DriveFileEntity, 0, len(files))
-	for _, f := range files {
-		out = append(out, h.packDriveFileAdmin(f))
-	}
-	return c.JSON(http.StatusOK, out)
-}
-
-// packDriveFileAdmin packs a drive file and embeds the owning user as
-// nested `user` when userRepo is wired. Folder is left nil because the
-// admin handler does not currently wire a DriveFolderRepository.
-func (h *Handler) packDriveFileAdmin(f *model.DriveFile) entity.DriveFileEntity {
-	// user repoが未設定なら user は nil fallback。folder は admin handler
-	// に folderRepoがないため常に nil (必要になれば拡張)。
-	var user *model.User
-	if h.userRepo != nil && f.UserID != nil {
-		if u, err := h.userRepo.FindByID(*f.UserID); err == nil {
-			user = u
-		}
-	}
-	return entity.PackDriveFileWithRelations(f, h.idGen, nil, user)
-}
-
-// DriveShowFile handles POST /api/admin/drive/show-file.
-// Accepts either a fileId or a url as identifier (Misskey 本家互換)。
-func (h *Handler) DriveShowFile(c echo.Context) error {
-	var req struct {
-		FileID string `json:"fileId"`
-		URL    string `json:"url"`
-	}
-	if err := c.Bind(&req); err != nil || (req.FileID == "" && req.URL == "") {
-		return c.JSON(http.StatusBadRequest, apierr.Error("INVALID_PARAM", "fileId or url is required.", "3d81ceae-475f-4600-b2a8-2bc116157532"))
-	}
-	if h.driveFileRepo == nil {
-		return c.JSON(http.StatusNotFound, apierr.Error("NO_SUCH_FILE", "No such file.", "ac4f7b11-1a6e-47e3-bf3d-3dce9a0e07ab"))
-	}
-	if req.FileID != "" {
-		file, err := h.driveFileRepo.FindByID(req.FileID)
-		if err != nil {
-			return c.JSON(http.StatusNotFound, apierr.Error("NO_SUCH_FILE", "No such file.", "ac4f7b11-1a6e-47e3-bf3d-3dce9a0e07ab"))
-		}
-		return c.JSON(http.StatusOK, h.packDriveFileAdmin(file))
-	}
-	// url 指定時は adminDB を使って url / thumbnailUrl / webpublicUrl いずれか
-	// に一致する 1 件を引く。 driveFileRepo には該当 API が無いため raw query で。
-	if h.adminDB == nil {
-		return c.JSON(http.StatusNotFound, apierr.Error("NO_SUCH_FILE", "No such file.", "ac4f7b11-1a6e-47e3-bf3d-3dce9a0e07ab"))
-	}
-	var file model.DriveFile
-	if err := h.adminDB.Where(
-		`"url" = ? OR "thumbnailUrl" = ? OR "webpublicUrl" = ?`,
-		req.URL, req.URL, req.URL,
-	).First(&file).Error; err != nil {
-		return c.JSON(http.StatusNotFound, apierr.Error("NO_SUCH_FILE", "No such file.", "ac4f7b11-1a6e-47e3-bf3d-3dce9a0e07ab"))
-	}
-	return c.JSON(http.StatusOK, h.packDriveFileAdmin(&file))
-}
-
-// --- emoji bulk ops ---
-
-// EmojiAddAliasesBulk handles POST /api/admin/emoji/add-aliases-bulk.
-//
-// 旧実装は id ごとに FindByID→UpdateFields を直列に呼んでいたが、件数が多いと
-// DB 往復が線形に増える。FindManyByIDs で一括取得 → 個別に alias を merge →
-// UpdateFields で書き戻す (alias はレコード毎に異なるため UpdateFieldsMany
-// では一括化できない)。
-func (h *Handler) EmojiAddAliasesBulk(c echo.Context) error {
-	var req struct {
-		IDs     []string `json:"ids"`
-		Aliases []string `json:"aliases"`
-	}
-	if err := c.Bind(&req); err != nil || len(req.IDs) == 0 {
-		return c.NoContent(http.StatusNoContent)
-	}
-	if h.emojiRepo == nil {
-		return c.NoContent(http.StatusNoContent)
-	}
-	rows, err := h.emojiRepo.FindManyByIDs(req.IDs)
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, apierr.Error("INTERNAL_ERROR", "Internal error.", "00000000-0000-0000-0000-000000000000"))
-	}
-	for _, e := range rows {
-		merged := dedupe(append(append([]string{}, e.Aliases...), req.Aliases...))
-		_ = h.emojiRepo.UpdateFields(e.ID, map[string]any{"aliases": merged})
-	}
-	return c.NoContent(http.StatusNoContent)
-}
-
-// EmojiCopy handles POST /api/admin/emoji/copy.
-//
-// Misskey 本家は対象 name が既存ならば DUPLICATE_NAME を返す。サフィックス
-// _copy を付けても既存に衝突する可能性があるため、明示チェックを入れる。
-func (h *Handler) EmojiCopy(c echo.Context) error {
-	var req struct {
-		EmojiID string `json:"emojiId"`
-	}
-	if err := c.Bind(&req); err != nil || req.EmojiID == "" {
-		return c.JSON(http.StatusBadRequest, apierr.Error("INVALID_PARAM", "emojiId is required.", "00000000-0000-0000-0000-000000000000"))
-	}
-	if h.emojiRepo == nil {
-		return c.NoContent(http.StatusNoContent)
-	}
-	src, err := h.emojiRepo.FindByID(req.EmojiID)
-	if err != nil {
-		return c.JSON(http.StatusNotFound, apierr.Error("NO_SUCH_EMOJI", "No such emoji.", "e2785b66-dca3-4087-9cac-b93c541cc425"))
-	}
-	newName := src.Name + "_copy"
-	if existing, err := h.emojiRepo.FindByNameAndHost(newName, nil); err == nil && existing != nil {
-		return c.JSON(http.StatusBadRequest, apierr.Error("DUPLICATE_NAME", "Duplicate name.", "5abef7f4-b3d6-4be4-a08f-4c4b7cf4f5b0"))
-	}
-	copied := *src
-	copied.ID = h.idGen.Generate(time.Now())
-	copied.Name = newName
-	copied.Host = nil
-	if err := h.emojiRepo.Create(&copied); err != nil {
-		return c.JSON(http.StatusInternalServerError, apierr.Error("INTERNAL_ERROR", "Internal error.", "00000000-0000-0000-0000-000000000000"))
-	}
-	return c.JSON(http.StatusOK, map[string]any{"id": copied.ID})
-}
-
-// EmojiDeleteBulk handles POST /api/admin/emoji/delete-bulk.
-func (h *Handler) EmojiDeleteBulk(c echo.Context) error {
-	var req struct {
-		IDs []string `json:"ids"`
-	}
-	if err := c.Bind(&req); err != nil || len(req.IDs) == 0 {
-		return c.NoContent(http.StatusNoContent)
-	}
-	if h.emojiRepo == nil {
-		return c.NoContent(http.StatusNoContent)
-	}
-	if err := h.emojiRepo.DeleteMany(req.IDs); err != nil {
-		return c.JSON(http.StatusInternalServerError, apierr.Error("INTERNAL_ERROR", "Internal error.", "00000000-0000-0000-0000-000000000000"))
-	}
-	return c.NoContent(http.StatusNoContent)
-}
-
-// dedupe returns the slice with duplicates removed while preserving order.
-func dedupe(in []string) []string {
-	seen := make(map[string]bool, len(in))
-	out := make([]string, 0, len(in))
-	for _, s := range in {
-		if seen[s] {
-			continue
-		}
-		seen[s] = true
-		out = append(out, s)
-	}
-	return out
-}
-
-// EmojiImportZip handles POST /api/admin/emoji/import-zip.
-//
-// fileId で Drive にアップロード済みの ZIP を指定し、非同期ジョブで展開して
-// ローカルカスタム絵文字として登録する。本家 Misskey 互換
-// (QueueService.createImportCustomEmojisJob 相当)。
-func (h *Handler) EmojiImportZip(c echo.Context) error {
-	var req struct {
-		FileID string `json:"fileId"`
-	}
-	if err := c.Bind(&req); err != nil || req.FileID == "" {
-		return c.JSON(http.StatusBadRequest, apierr.Error("INVALID_PARAM", "fileId is required.", "5f4c9d8a-7c39-4bfa-9dcb-09f17e0f7a25"))
-	}
-	// drive file の存在確認 (ジョブが後で失敗するよりここで早期に弾く)
-	if h.driveFileRepo != nil {
-		if _, err := h.driveFileRepo.FindByID(req.FileID); err != nil {
-			return c.JSON(http.StatusBadRequest, apierr.Error("NO_SUCH_FILE", "No such file.", "ac4f7b11-1a6e-47e3-bf3d-3dce9a0e07ab"))
-		}
-	}
-	if h.emojiEnqueuer == nil {
-		return c.NoContent(http.StatusNoContent)
-	}
-	user := middleware.GetUser(c)
-	if user == nil {
-		return c.NoContent(http.StatusNoContent)
-	}
-	if err := h.emojiEnqueuer.EnqueueImportCustomEmojis(queue.ImportCustomEmojisPayload{
-		UserID: user.ID,
-		FileID: req.FileID,
-	}); err != nil {
-		return c.JSON(http.StatusInternalServerError, apierr.Error("INTERNAL_ERROR", "Failed to enqueue emoji import.", "89a6d9fd-0fe6-4c3c-9daa-7c6b1f29f1a4"))
-	}
-	return c.NoContent(http.StatusNoContent)
-}
-
-// EmojiListRemote handles POST /api/admin/emoji/list-remote.
-func (h *Handler) EmojiListRemote(c echo.Context) error {
-	if h.emojiRepo == nil {
-		return c.JSON(http.StatusOK, []any{})
-	}
-	var req struct {
-		Query   string `json:"query"`
-		Host    string `json:"host"`
-		SinceID string `json:"sinceId"`
-		UntilID string `json:"untilId"`
-		Limit   int    `json:"limit"`
-		Offset  int    `json:"offset"`
-	}
-	_ = c.Bind(&req)
-	if req.Limit <= 0 || req.Limit > 100 {
-		req.Limit = 30
-	}
-	emojis, err := h.emojiRepo.ListRemoteWithFilter(req.Query, req.Host, req.SinceID, req.UntilID, req.Limit, req.Offset)
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, apierr.Error("INTERNAL_ERROR", "Internal error.", "00000000-0000-0000-0000-000000000000"))
-	}
-	if emojis == nil {
-		return c.JSON(http.StatusOK, []any{})
-	}
-	// upstream Misskey の admin/emoji/list-remote は EmojiDetailed schema を
-	// 返す (= `url` field、`publicUrl` ではない)。frontend の旧
-	// custom-emojis-manager.vue は emoji.url を読み込んで <img> を組み立てる
-	// ため、raw model.Emoji (publicUrl / originalUrl 持ち、url 無し) を返すと
-	// 画像が壊れて表示そのものが空に見える (#466)。entity.PackEmojiDetailedList
-	// で publicUrl → url 変換した shape にする。
-	return c.JSON(http.StatusOK, entity.PackEmojiDetailedList(emojis))
-}
-
-// EmojiRemoveAliasesBulk handles POST /api/admin/emoji/remove-aliases-bulk.
-func (h *Handler) EmojiRemoveAliasesBulk(c echo.Context) error {
-	var req struct {
-		IDs     []string `json:"ids"`
-		Aliases []string `json:"aliases"`
-	}
-	if err := c.Bind(&req); err != nil || len(req.IDs) == 0 {
-		return c.NoContent(http.StatusNoContent)
-	}
-	if h.emojiRepo == nil {
-		return c.NoContent(http.StatusNoContent)
-	}
-	rows, err := h.emojiRepo.FindManyByIDs(req.IDs)
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, apierr.Error("INTERNAL_ERROR", "Internal error.", "00000000-0000-0000-0000-000000000000"))
-	}
-	removeSet := make(map[string]bool, len(req.Aliases))
-	for _, a := range req.Aliases {
-		removeSet[a] = true
-	}
-	for _, e := range rows {
-		filtered := make([]string, 0, len(e.Aliases))
-		for _, a := range e.Aliases {
-			if !removeSet[a] {
-				filtered = append(filtered, a)
-			}
-		}
-		_ = h.emojiRepo.UpdateFields(e.ID, map[string]any{"aliases": filtered})
-	}
-	return c.NoContent(http.StatusNoContent)
-}
-
-// EmojiSetAliasesBulk handles POST /api/admin/emoji/set-aliases-bulk.
-//
-// 全 id で同じ aliases 値を設定するため UpdateFieldsMany (WHERE IN UPDATE) で
-// 1 文に集約できる。
-func (h *Handler) EmojiSetAliasesBulk(c echo.Context) error {
-	var req struct {
-		IDs     []string `json:"ids"`
-		Aliases []string `json:"aliases"`
-	}
-	if err := c.Bind(&req); err != nil || len(req.IDs) == 0 {
-		return c.NoContent(http.StatusNoContent)
-	}
-	if h.emojiRepo == nil {
-		return c.NoContent(http.StatusNoContent)
-	}
-	if err := h.emojiRepo.UpdateFieldsMany(req.IDs, map[string]any{"aliases": req.Aliases}); err != nil {
-		return c.JSON(http.StatusInternalServerError, apierr.Error("INTERNAL_ERROR", "Internal error.", "00000000-0000-0000-0000-000000000000"))
-	}
-	return c.NoContent(http.StatusNoContent)
-}
-
-// EmojiSetCategoryBulk handles POST /api/admin/emoji/set-category-bulk.
-func (h *Handler) EmojiSetCategoryBulk(c echo.Context) error {
-	var req struct {
-		IDs      []string `json:"ids"`
-		Category string   `json:"category"`
-	}
-	if err := c.Bind(&req); err != nil || len(req.IDs) == 0 {
-		return c.NoContent(http.StatusNoContent)
-	}
-	if h.emojiRepo == nil {
-		return c.NoContent(http.StatusNoContent)
-	}
-	if err := h.emojiRepo.UpdateFieldsMany(req.IDs, map[string]any{"category": req.Category}); err != nil {
-		return c.JSON(http.StatusInternalServerError, apierr.Error("INTERNAL_ERROR", "Internal error.", "00000000-0000-0000-0000-000000000000"))
-	}
-	return c.NoContent(http.StatusNoContent)
-}
-
-// EmojiSetLicenseBulk handles POST /api/admin/emoji/set-license-bulk.
-func (h *Handler) EmojiSetLicenseBulk(c echo.Context) error {
-	var req struct {
-		IDs     []string `json:"ids"`
-		License string   `json:"license"`
-	}
-	if err := c.Bind(&req); err != nil || len(req.IDs) == 0 {
-		return c.NoContent(http.StatusNoContent)
-	}
-	if h.emojiRepo == nil {
-		return c.NoContent(http.StatusNoContent)
-	}
-	if err := h.emojiRepo.UpdateFieldsMany(req.IDs, map[string]any{"license": req.License}); err != nil {
-		return c.JSON(http.StatusInternalServerError, apierr.Error("INTERNAL_ERROR", "Internal error.", "00000000-0000-0000-0000-000000000000"))
-	}
-	return c.NoContent(http.StatusNoContent)
-}
-
-// --- captcha ---
-
-// CaptchaCurrent handles POST /api/admin/captcha/current.
-//
-// Returns which captcha provider (if any) is currently enabled and its public
-// site key so the admin UI can render the correct configuration form.
-func (h *Handler) CaptchaCurrent(c echo.Context) error {
-	if h.metaRepo == nil {
-		return c.JSON(http.StatusOK, map[string]any{"provider": nil})
-	}
-	meta, err := h.metaRepo.Fetch()
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, apierr.Error("INTERNAL_ERROR", "Internal error.", "00000000-0000-0000-0000-000000000000"))
-	}
-	var provider any
-	var siteKey *string
-	switch {
-	case meta.EnableHcaptcha:
-		provider = "hcaptcha"
-		siteKey = meta.HcaptchaSiteKey
-	case meta.EnableRecaptcha:
-		provider = "recaptcha"
-		siteKey = meta.RecaptchaSiteKey
-	case meta.EnableTurnstile:
-		provider = "turnstile"
-		siteKey = meta.TurnstileSiteKey
-	}
-	return c.JSON(http.StatusOK, map[string]any{
-		"provider": provider,
-		"siteKey":  siteKey,
-	})
-}
-
-// CaptchaSave handles POST /api/admin/captcha/save.
-func (h *Handler) CaptchaSave(c echo.Context) error {
-	var req struct {
-		Provider    string  `json:"provider"`
-		HcaptchaSK  *string `json:"hcaptchaSiteKey"`
-		HcaptchaSS  *string `json:"hcaptchaSecretKey"`
-		RecaptchaSK *string `json:"recaptchaSiteKey"`
-		RecaptchaSS *string `json:"recaptchaSecretKey"`
-		TurnstileSK *string `json:"turnstileSiteKey"`
-		TurnstileSS *string `json:"turnstileSecretKey"`
-	}
-	if err := c.Bind(&req); err != nil {
-		return c.JSON(http.StatusBadRequest, apierr.Error("INVALID_PARAM", "Invalid parameters.", "00000000-0000-0000-0000-000000000000"))
-	}
-	// provider が空 or "none" の場合は captcha 無効化として扱う。
-	switch req.Provider {
-	case "", "none", "hcaptcha", "recaptcha", "turnstile":
-	default:
-		return c.JSON(http.StatusBadRequest, apierr.Error("INVALID_PARAM", "Unknown captcha provider.", "00000000-0000-0000-0000-000000000000"))
-	}
-	if h.metaRepo == nil {
-		return c.NoContent(http.StatusNoContent)
-	}
-	fields := map[string]any{
-		"enableHcaptcha":  req.Provider == "hcaptcha",
-		"enableRecaptcha": req.Provider == "recaptcha",
-		"enableTurnstile": req.Provider == "turnstile",
-	}
-	if req.HcaptchaSK != nil {
-		fields["hcaptchaSiteKey"] = *req.HcaptchaSK
-	}
-	if req.HcaptchaSS != nil {
-		fields["hcaptchaSecretKey"] = *req.HcaptchaSS
-	}
-	if req.RecaptchaSK != nil {
-		fields["recaptchaSiteKey"] = *req.RecaptchaSK
-	}
-	if req.RecaptchaSS != nil {
-		fields["recaptchaSecretKey"] = *req.RecaptchaSS
-	}
-	if req.TurnstileSK != nil {
-		fields["turnstileSiteKey"] = *req.TurnstileSK
-	}
-	if req.TurnstileSS != nil {
-		fields["turnstileSecretKey"] = *req.TurnstileSS
-	}
-	if err := h.metaRepo.Update(fields); err != nil {
-		return c.JSON(http.StatusInternalServerError, apierr.Error("INTERNAL_ERROR", "Internal error.", "00000000-0000-0000-0000-000000000000"))
-	}
-	return c.NoContent(http.StatusNoContent)
-}
-
-// --- ad ---
-
-// AdCreate handles POST /api/admin/ad/create.
-func (h *Handler) AdCreate(c echo.Context) error {
-	if h.adRepo == nil {
-		return c.NoContent(http.StatusNoContent)
-	}
-	var req struct {
-		URL         string `json:"url"`
-		ImageURL    string `json:"imageUrl"`
-		Place       string `json:"place"`
-		Memo        string `json:"memo"`
-		Priority    string `json:"priority"`
-		Ratio       int    `json:"ratio"`
-		DayOfWeek   int    `json:"dayOfWeek"`
-		ExpiresAt   int64  `json:"expiresAt"`
-		StartsAt    int64  `json:"startsAt"`
-		IsSensitive bool   `json:"isSensitive"`
-	}
-	if err := c.Bind(&req); err != nil || req.URL == "" || req.ImageURL == "" {
-		return c.JSON(http.StatusBadRequest, apierr.Error("INVALID_PARAM", "Invalid parameters.", "00000000-0000-0000-0000-000000000000"))
-	}
-	if req.Place == "" {
-		req.Place = "square"
-	}
-	if req.Priority == "" {
-		req.Priority = "middle"
-	}
-	if req.Ratio <= 0 {
-		req.Ratio = 1
-	}
-	ad := &model.Ad{
-		ID:          h.idGen.Generate(time.Now()),
-		URL:         req.URL,
-		ImageURL:    req.ImageURL,
-		Place:       req.Place,
-		Memo:        req.Memo,
-		Priority:    req.Priority,
-		Ratio:       req.Ratio,
-		DayOfWeek:   req.DayOfWeek,
-		IsSensitive: req.IsSensitive,
-		StartsAt:    millisOrNow(req.StartsAt),
-		ExpiresAt:   millisOrDefault(req.ExpiresAt, time.Now().Add(30*24*time.Hour)),
-	}
-	if err := h.adRepo.Create(ad); err != nil {
-		return c.JSON(http.StatusInternalServerError, apierr.Error("INTERNAL_ERROR", "Internal error.", "00000000-0000-0000-0000-000000000000"))
-	}
-	return c.JSON(http.StatusOK, ad)
-}
-
-// AdDelete handles POST /api/admin/ad/delete.
-func (h *Handler) AdDelete(c echo.Context) error {
-	if h.adRepo == nil {
-		return c.NoContent(http.StatusNoContent)
-	}
-	var req struct {
-		ID string `json:"id"`
-	}
-	_ = c.Bind(&req)
-	_ = h.adRepo.Delete(req.ID)
-	return c.NoContent(http.StatusNoContent)
-}
-
-// AdList handles POST /api/admin/ad/list.
-func (h *Handler) AdList(c echo.Context) error {
-	if h.adRepo == nil {
-		return c.JSON(http.StatusOK, []any{})
-	}
-	var req struct {
-		Limit  int `json:"limit"`
-		Offset int `json:"offset"`
-	}
-	_ = c.Bind(&req)
-	if req.Limit <= 0 || req.Limit > 100 {
-		req.Limit = 30
-	}
-	rows, err := h.adRepo.List(req.Limit, req.Offset)
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, apierr.Error("INTERNAL_ERROR", "Internal error.", "00000000-0000-0000-0000-000000000000"))
-	}
-	if rows == nil {
-		return c.JSON(http.StatusOK, []any{})
-	}
-	return c.JSON(http.StatusOK, rows)
-}
-
-// AdUpdate handles POST /api/admin/ad/update.
-func (h *Handler) AdUpdate(c echo.Context) error {
-	// 旧実装が `c.Request().Body` を `Updates` に直接渡しており部分更新が機能して
-	// いなかった。partial field map に差し替え、リクエストで明示された項目のみ
-	// 書き換える。
-	if h.adRepo == nil {
-		return c.NoContent(http.StatusNoContent)
-	}
-	var req struct {
-		ID          string  `json:"id"`
-		URL         *string `json:"url"`
-		ImageURL    *string `json:"imageUrl"`
-		Place       *string `json:"place"`
-		Memo        *string `json:"memo"`
-		Priority    *string `json:"priority"`
-		Ratio       *int    `json:"ratio"`
-		DayOfWeek   *int    `json:"dayOfWeek"`
-		ExpiresAt   *int64  `json:"expiresAt"`
-		StartsAt    *int64  `json:"startsAt"`
-		IsSensitive *bool   `json:"isSensitive"`
-	}
-	if err := c.Bind(&req); err != nil || req.ID == "" {
-		return c.JSON(http.StatusBadRequest, apierr.Error("INVALID_PARAM", "Invalid parameters.", "00000000-0000-0000-0000-000000000000"))
-	}
-	if _, err := h.adRepo.FindByID(req.ID); err != nil {
-		return c.JSON(http.StatusNotFound, apierr.Error("NOT_FOUND", "Not found.", "00000000-0000-0000-0000-000000000000"))
-	}
-	fields := map[string]any{}
-	if req.URL != nil {
-		fields["url"] = *req.URL
-	}
-	if req.ImageURL != nil {
-		fields["imageUrl"] = *req.ImageURL
-	}
-	if req.Place != nil {
-		fields["place"] = *req.Place
-	}
-	if req.Memo != nil {
-		fields["memo"] = *req.Memo
-	}
-	if req.Priority != nil {
-		fields["priority"] = *req.Priority
-	}
-	if req.Ratio != nil {
-		fields["ratio"] = *req.Ratio
-	}
-	if req.DayOfWeek != nil {
-		fields["dayOfWeek"] = *req.DayOfWeek
-	}
-	if req.IsSensitive != nil {
-		fields["isSensitive"] = *req.IsSensitive
-	}
-	// pointer 型フィールドは nil (未指定) と 0 (明示的に 0) を区別できるので、
-	// 受け取った値をそのまま UnixMilli に変換する。Create 側と違い 0 を now に
-	// 読み替えない。
-	if req.StartsAt != nil {
-		fields["startsAt"] = time.UnixMilli(*req.StartsAt)
-	}
-	if req.ExpiresAt != nil {
-		fields["expiresAt"] = time.UnixMilli(*req.ExpiresAt)
-	}
-	if err := h.adRepo.UpdateFields(req.ID, fields); err != nil {
-		return c.JSON(http.StatusInternalServerError, apierr.Error("INTERNAL_ERROR", "Internal error.", "00000000-0000-0000-0000-000000000000"))
-	}
-	return c.NoContent(http.StatusNoContent)
-}
-
-// millisOrNow converts a UNIX millisecond timestamp to time.Time; 0 falls back
-// to the current wall clock.
-func millisOrNow(ms int64) time.Time {
-	if ms == 0 {
-		return time.Now()
-	}
-	return time.UnixMilli(ms)
-}
-
-// millisOrDefault converts a UNIX millisecond timestamp to time.Time; 0 falls
-// back to the provided default.
-func millisOrDefault(ms int64, def time.Time) time.Time {
-	if ms == 0 {
-		return def
-	}
-	return time.UnixMilli(ms)
-}
-
-// --- avatar-decorations ---
-
-// AvatarDecorationsCreate handles POST /api/admin/avatar-decorations/create.
-func (h *Handler) AvatarDecorationsCreate(c echo.Context) error {
-	if h.avatarDecoRepo == nil {
-		return c.NoContent(http.StatusNoContent)
-	}
-	var req struct {
-		Name        string   `json:"name"`
-		Description string   `json:"description"`
-		URL         string   `json:"url"`
-		RoleIDs     []string `json:"roleIdsThatCanBeUsedThisDecoration"`
-	}
-	if err := c.Bind(&req); err != nil || req.Name == "" || req.URL == "" {
-		return c.JSON(http.StatusBadRequest, apierr.Error("INVALID_PARAM", "Invalid parameters.", "00000000-0000-0000-0000-000000000000"))
-	}
-	d := &model.AvatarDecoration{
-		ID:          h.idGen.Generate(time.Now()),
-		Name:        req.Name,
-		Description: req.Description,
-		URL:         req.URL,
-		RoleIDs:     req.RoleIDs,
-	}
-	if err := h.avatarDecoRepo.Create(d); err != nil {
-		return c.JSON(http.StatusInternalServerError, apierr.Error("INTERNAL_ERROR", "Internal error.", "00000000-0000-0000-0000-000000000000"))
-	}
-	return c.JSON(http.StatusOK, d)
-}
-
-// AvatarDecorationsDelete handles POST /api/admin/avatar-decorations/delete.
-func (h *Handler) AvatarDecorationsDelete(c echo.Context) error {
-	if h.avatarDecoRepo == nil {
-		return c.NoContent(http.StatusNoContent)
-	}
-	var req struct {
-		ID string `json:"id"`
-	}
-	_ = c.Bind(&req)
-	_ = h.avatarDecoRepo.Delete(req.ID)
-	return c.NoContent(http.StatusNoContent)
-}
-
-// AvatarDecorationsList handles POST /api/admin/avatar-decorations/list.
-func (h *Handler) AvatarDecorationsList(c echo.Context) error {
-	if h.avatarDecoRepo == nil {
-		return c.JSON(http.StatusOK, []any{})
-	}
-	rows, err := h.avatarDecoRepo.List()
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, apierr.Error("INTERNAL_ERROR", "Internal error.", "00000000-0000-0000-0000-000000000000"))
-	}
-	if rows == nil {
-		return c.JSON(http.StatusOK, []any{})
-	}
-	return c.JSON(http.StatusOK, rows)
-}
-
-// AvatarDecorationsUpdate handles POST /api/admin/avatar-decorations/update.
-func (h *Handler) AvatarDecorationsUpdate(c echo.Context) error {
-	if h.avatarDecoRepo == nil {
-		return c.NoContent(http.StatusNoContent)
-	}
-	var req struct {
-		ID          string    `json:"id"`
-		Name        *string   `json:"name"`
-		Description *string   `json:"description"`
-		URL         *string   `json:"url"`
-		RoleIDs     *[]string `json:"roleIdsThatCanBeUsedThisDecoration"`
-	}
-	if err := c.Bind(&req); err != nil || req.ID == "" {
-		return c.JSON(http.StatusBadRequest, apierr.Error("INVALID_PARAM", "Invalid parameters.", "00000000-0000-0000-0000-000000000000"))
-	}
-	if _, err := h.avatarDecoRepo.FindByID(req.ID); err != nil {
-		return c.JSON(http.StatusNotFound, apierr.Error("NOT_FOUND", "Not found.", "00000000-0000-0000-0000-000000000000"))
-	}
-	fields := map[string]any{"updatedAt": time.Now()}
-	if req.Name != nil {
-		fields["name"] = *req.Name
-	}
-	if req.Description != nil {
-		fields["description"] = *req.Description
-	}
-	if req.URL != nil {
-		fields["url"] = *req.URL
-	}
-	if req.RoleIDs != nil {
-		fields["roleIdsThatCanBeUsedThisDecoration"] = *req.RoleIDs
-	}
-	if err := h.avatarDecoRepo.UpdateFields(req.ID, fields); err != nil {
-		return c.JSON(http.StatusInternalServerError, apierr.Error("INTERNAL_ERROR", "Internal error.", "00000000-0000-0000-0000-000000000000"))
-	}
-	return c.NoContent(http.StatusNoContent)
-}
-
-// --- federation ---
-
-// FederationDeleteAllFiles handles POST /api/admin/federation/delete-all-files.
-func (h *Handler) FederationDeleteAllFiles(c echo.Context) error {
-	var req struct {
-		Host string `json:"host"`
-	}
-	_ = c.Bind(&req)
-	// リモートファイル削除は将来対応 (DriveFileRepoのリモートファイル一括削除が必要)
-	return c.NoContent(http.StatusNoContent)
-}
-
-// FederationRefreshRemoteInstanceMetadata handles POST /api/admin/federation/refresh-remote-instance-metadata.
-// InstanceMetadataFetcher (= coreinstance.FetchMetadataService) 経由で指定ホストの
-// nodeinfo + iconUrl / faviconUrl を再取得する。fetcher 未設定または host 未指定
-// の場合は no-op で 204 を返す (本家 TS も失敗時エラーコードは返さない挙動)。
-func (h *Handler) FederationRefreshRemoteInstanceMetadata(c echo.Context) error {
-	var req struct {
-		Host string `json:"host"`
-	}
-	_ = c.Bind(&req)
-	if h.instanceMetadataFetcher == nil || req.Host == "" {
-		return c.NoContent(http.StatusNoContent)
-	}
-	// fetch 失敗はユーザーへ明示的にエラー返す必要はない (frontendは成功前提
-	// でUI更新するだけ)。ログに残してリトライ可能な状態にしておく。
-	if err := h.instanceMetadataFetcher.Fetch(req.Host); err != nil {
-		slog.Warn("federation refresh metadata failed", "host", req.Host, "err", err)
-	}
-	return c.NoContent(http.StatusNoContent)
-}
-
-// FederationRemoveAllFollowing handles POST /api/admin/federation/remove-all-following.
-func (h *Handler) FederationRemoveAllFollowing(c echo.Context) error {
-	var req struct {
-		Host string `json:"host"`
-	}
-	_ = c.Bind(&req)
-	// リモートフォロー一括削除は将来対応
-	return c.NoContent(http.StatusNoContent)
-}
-
-// FederationUpdateInstance handles POST /api/admin/federation/update-instance.
-func (h *Handler) FederationUpdateInstance(c echo.Context) error {
-	if h.adminDB == nil {
-		return c.NoContent(http.StatusNoContent)
-	}
-	var req struct {
-		Host           string `json:"host"`
-		IsSuspended    *bool  `json:"isSuspended"`
-		IsBlocked      *bool  `json:"isBlocked"`
-		IsSilenced     *bool  `json:"isSilenced"`
-		ModerationNote string `json:"moderationNote"`
-	}
-	if err := c.Bind(&req); err != nil || req.Host == "" {
-		return c.NoContent(http.StatusNoContent)
-	}
-	updates := map[string]any{}
-	if req.IsSuspended != nil {
-		updates["isSuspended"] = *req.IsSuspended
-	}
-	if req.IsBlocked != nil {
-		updates["isBlocked"] = *req.IsBlocked
-	}
-	if req.IsSilenced != nil {
-		updates["isSilenced"] = *req.IsSilenced
-	}
-	if req.ModerationNote != "" {
-		updates["moderationNote"] = req.ModerationNote
-	}
-	if len(updates) > 0 {
-		h.adminDB.Model(&model.Instance{}).Where(`"host" = ?`, req.Host).Updates(updates)
-	}
-	return c.NoContent(http.StatusNoContent)
-}
-
-// --- invite ---
-
-// InviteCreate handles POST /api/admin/invite/create.
-func (h *Handler) InviteCreate(c echo.Context) error {
-	// 本家 TS は count (1-100, default 1) 分のチケットを配列で返す。個々の Create
-	// 失敗時でも既作成分はロールバックしない (本家も Promise.all で非原子的)。
-	if h.inviteRepo == nil {
-		return c.NoContent(http.StatusNoContent)
-	}
-	var req struct {
-		Count     int     `json:"count"`
-		ExpiresAt *string `json:"expiresAt"`
-	}
-	_ = c.Bind(&req)
-	if req.Count <= 0 {
-		req.Count = 1
-	}
-	if req.Count > 100 {
-		req.Count = 100
-	}
-	var expiresAt *time.Time
-	if req.ExpiresAt != nil && *req.ExpiresAt != "" {
-		parsed, err := time.Parse(time.RFC3339, *req.ExpiresAt)
-		if err != nil {
-			return c.JSON(http.StatusBadRequest, apierr.Error("INVALID_DATE_TIME", "Invalid date-time format", "f1380b15-3760-4c6c-a1db-5c3aaf1cbd49"))
-		}
-		expiresAt = &parsed
-	}
-	user := middleware.GetUser(c)
-	var createdByID *string
-	if user != nil {
-		createdByID = &user.ID
-	}
-	tickets := make([]*model.RegistrationTicket, 0, req.Count)
-	for i := 0; i < req.Count; i++ {
-		b := make([]byte, 16)
-		_, _ = rand.Read(b)
-		t := &model.RegistrationTicket{
-			ID:          h.idGen.Generate(time.Now()),
-			Code:        hex.EncodeToString(b),
-			ExpiresAt:   expiresAt,
-			CreatedByID: createdByID,
-		}
-		if err := h.inviteRepo.Create(t); err != nil {
-			continue
-		}
-		tickets = append(tickets, t)
-	}
-	return c.JSON(http.StatusOK, h.packInviteTickets(tickets))
-}
-
-// packInviteTickets transforms RegistrationTicket rows into the
-// Misskey-compatible InviteCodeEntityService.pack shape.
-func (h *Handler) packInviteTickets(rows []*model.RegistrationTicket) []map[string]any {
-	// Misskey 本家 InviteCodeEntityService.pack と同じ形にする。
-	// createdAt は aidx ID から抽出、used は usedAt の有無で導出する。
-	out := make([]map[string]any, 0, len(rows))
-	for _, t := range rows {
-		var createdAt *string
-		if h.idGen != nil {
-			if ts, err := h.idGen.ParseTime(t.ID); err == nil {
-				s := ts.UTC().Format("2006-01-02T15:04:05.000Z")
-				createdAt = &s
-			}
-		}
-		var expiresAt *string
-		if t.ExpiresAt != nil {
-			s := t.ExpiresAt.UTC().Format("2006-01-02T15:04:05.000Z")
-			expiresAt = &s
-		}
-		var usedAt *string
-		if t.UsedAt != nil {
-			s := t.UsedAt.UTC().Format("2006-01-02T15:04:05.000Z")
-			usedAt = &s
-		}
-		out = append(out, map[string]any{
-			"id":          t.ID,
-			"code":        t.Code,
-			"expiresAt":   expiresAt,
-			"createdAt":   createdAt,
-			"createdBy":   nil,
-			"usedBy":      nil,
-			"usedAt":      usedAt,
-			"used":        t.UsedAt != nil,
-			"createdById": t.CreatedByID,
-			"usedById":    t.UsedByID,
-		})
-	}
-	return out
-}
-
-// InviteList handles POST /api/admin/invite/list.
-func (h *Handler) InviteList(c echo.Context) error {
-	if h.inviteRepo == nil {
-		return c.JSON(http.StatusOK, []any{})
-	}
-	var req struct {
-		Limit  int    `json:"limit"`
-		Offset int    `json:"offset"`
-		Type   string `json:"type"`
-	}
-	_ = c.Bind(&req)
-	if req.Limit <= 0 || req.Limit > 100 {
-		req.Limit = 30
-	}
-	filter := req.Type
-	switch filter {
-	case "unused", "used", "expired", "all":
-	default:
-		filter = "all"
-	}
-	rows, err := h.inviteRepo.List(filter, req.Limit, req.Offset, time.Now())
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, apierr.Error("INTERNAL_ERROR", "Internal error.", "00000000-0000-0000-0000-000000000000"))
-	}
-	return c.JSON(http.StatusOK, h.packInviteTickets(rows))
-}
-
-// --- promo ---
-
-// PromoCreate handles POST /api/admin/promo/create.
-func (h *Handler) PromoCreate(c echo.Context) error {
-	if h.promoNoteRepo == nil {
-		return c.NoContent(http.StatusNoContent)
-	}
-	var req struct {
-		NoteID    string `json:"noteId"`
-		ExpiresAt int64  `json:"expiresAt"`
-	}
-	if err := c.Bind(&req); err != nil || req.NoteID == "" {
-		return c.JSON(http.StatusBadRequest, apierr.Error("INVALID_PARAM", "noteId is required.", "3d81ceae-475f-4600-b2a8-2bc116157532"))
-	}
-
-	// 対象 note の存在確認 → 既に promote 済みでないか確認
-	var targetUserID string
-	if h.noteFinder != nil {
-		note, err := h.noteFinder.FindByID(req.NoteID)
-		if err != nil {
-			return c.JSON(http.StatusBadRequest, apierr.Error("NO_SUCH_NOTE", "No such note.", "ee449fbe-af2a-453b-9cae-cf2fe7c895fc"))
-		}
-		targetUserID = note.UserID
-	}
-
-	exists, err := h.promoNoteRepo.Exists(req.NoteID)
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, apierr.Error("INTERNAL_ERROR", "Internal error.", "00000000-0000-0000-0000-000000000000"))
-	}
-	if exists {
-		return c.JSON(http.StatusBadRequest, apierr.Error("ALREADY_PROMOTED", "The note has already promoted.", "ae427aa2-7a41-484f-a18c-2c1104051604"))
-	}
-
-	// note 情報が取れなかった場合は呼び出し admin を userId として記録 (最低限の整合)
-	if targetUserID == "" {
-		if u := middleware.GetUser(c); u != nil {
-			targetUserID = u.ID
-		}
-	}
-
-	promo := &model.PromoNote{
-		NoteID:    req.NoteID,
-		ExpiresAt: time.UnixMilli(req.ExpiresAt),
-		UserID:    targetUserID,
-	}
-	if err := h.promoNoteRepo.Create(promo); err != nil {
-		return c.JSON(http.StatusInternalServerError, apierr.Error("INTERNAL_ERROR", "Internal error.", "00000000-0000-0000-0000-000000000000"))
-	}
-	return c.NoContent(http.StatusNoContent)
-}
-
-// --- relays ---
-
-// RelaysAdd handles POST /api/admin/relays/add. Routes through the
-// relay Service so that a Follow activity is actually dispatched to
-// the relay's inbox (#161). Falls back to raw DB insertion when the
-// service is not wired (tests or early boot).
-func (h *Handler) RelaysAdd(c echo.Context) error {
-	var req struct {
-		Inbox string `json:"inbox"`
-	}
-	_ = c.Bind(&req)
-	if req.Inbox == "" {
-		return c.NoContent(http.StatusNoContent)
-	}
-	if h.relayService != nil {
-		rel, err := h.relayService.Add(c.Request().Context(), req.Inbox)
-		if err != nil {
-			return apierr.JSONInternalError(c)
-		}
-		return c.JSON(http.StatusOK, rel)
-	}
-	if h.adminDB == nil {
-		return c.NoContent(http.StatusNoContent)
-	}
-	relay := &model.Relay{
-		ID: h.idGen.Generate(time.Now()), Inbox: req.Inbox, Status: "requesting",
-	}
-	h.adminDB.Create(relay)
-	return c.JSON(http.StatusOK, relay)
-}
-
-// RelaysList handles POST /api/admin/relays/list.
-func (h *Handler) RelaysList(c echo.Context) error {
-	if h.relayService != nil {
-		list, err := h.relayService.List(c.Request().Context())
-		if err != nil {
-			return apierr.JSONInternalError(c)
-		}
-		if list == nil {
-			list = []*model.Relay{}
-		}
-		return c.JSON(http.StatusOK, list)
-	}
-	if h.adminDB == nil {
-		return c.JSON(http.StatusOK, []any{})
-	}
-	var relays []*model.Relay
-	h.adminDB.Order(`"id" DESC`).Find(&relays)
-	return c.JSON(http.StatusOK, relays)
-}
-
-// RelaysRemove handles POST /api/admin/relays/remove. When the relay
-// service is configured, it sends an Undo(Follow) to the relay inbox
-// before deleting the row.
-func (h *Handler) RelaysRemove(c echo.Context) error {
-	var req struct {
-		ID    string `json:"id"`
-		Inbox string `json:"inbox"`
-	}
-	_ = c.Bind(&req)
-	if h.relayService != nil {
-		id := req.ID
-		if id == "" && req.Inbox != "" && h.adminDB != nil {
-			// 本家互換のため inbox 指定でも remove できるようにする
-			var rel model.Relay
-			if err := h.adminDB.Where(`"inbox" = ?`, req.Inbox).First(&rel).Error; err == nil {
-				id = rel.ID
-			}
-		}
-		if id == "" {
-			return c.NoContent(http.StatusNoContent)
-		}
-		if err := h.relayService.Remove(c.Request().Context(), id); err != nil {
-			return apierr.JSONInternalError(c)
-		}
-		return c.NoContent(http.StatusNoContent)
-	}
-	if h.adminDB == nil {
-		return c.NoContent(http.StatusNoContent)
-	}
-	h.adminDB.Where(`"id" = ?`, req.ID).Delete(&model.Relay{})
-	return c.NoContent(http.StatusNoContent)
-}
