+++ internal/api/admin/proxy_account.go
@@ -1,0 +1,40 @@
+package admin
+
+import (
+	"net/http"
+
+	"github.com/labstack/echo/v4"
+	"github.com/shiroha-a/mk/internal/api/apierr"
+	"github.com/shiroha-a/mk/internal/entity"
+)
+
+// UpdateProxyAccount handles POST /api/admin/update-proxy-account.
+//
+// 本家 update-proxy-account.ts と同じく proxy system user の profile.description
+// を更新する handler (旧 stub は誤って meta.proxyAccountId を書き換えていた、#348)。
+// frontend admin/settings 画面の proxyAccountForm が `{ description }` 形式で
+// 呼び出すのでこれに合わせる。Response は UserDetailed (mk-go では UserLite +
+// description で代替)。
+func (h *Handler) UpdateProxyAccount(c echo.Context) error {
+	var req struct {
+		Description *string `json:"description"`
+	}
+	if err := c.Bind(&req); err != nil {
+		return c.JSON(http.StatusBadRequest, apierr.Error("INVALID_PARAM", "Invalid parameters.", "00000000-0000-0000-0000-000000000000"))
+	}
+	if h.systemAccountFetcher == nil || h.userRepo == nil {
+		return c.NoContent(http.StatusNoContent)
+	}
+	proxy, err := h.systemAccountFetcher.Fetch("proxy")
+	if err != nil || proxy == nil {
+		return c.JSON(http.StatusInternalServerError, apierr.Error("INTERNAL_ERROR", "Internal error.", "00000000-0000-0000-0000-000000000000"))
+	}
+	if req.Description != nil {
+		if err := h.userRepo.UpdateProfile(proxy.ID, map[string]any{"description": req.Description}); err != nil {
+			return c.JSON(http.StatusInternalServerError, apierr.Error("INTERNAL_ERROR", "Internal error.", "00000000-0000-0000-0000-000000000000"))
+		}
+	}
+	// 更新後 profile を再取得して UserDetailed を返す。
+	profile, _ := h.userRepo.FindProfileByUserID(proxy.ID)
+	return c.JSON(http.StatusOK, entity.PackUserDetailed(proxy, profile))
+}
