+++ internal/api/admin/proxy_account_test.go
@@ -1,0 +1,4 @@
+package admin_test
+
+// proxy account 系の詳細テストは accounts_test.go に常駐 (TestUpdateProxyAccount_*)。
+// 本ファイルは将来 proxy 関連のスモーク test を追加するときの placeholder。
